Added the command field and update key components in real time
