+++ src/map/include/entitySystem.h
@@ -10,6 +10,7 @@ class EntitySystem {
         EntitySystem() : systemManager_(entityManager_) {
             // TODO : use on_component_removed for Destination
             systemManager_.add<MovementSystem>();
+            systemManager_.add<UpdateComponents>();
         }
 
         void update(double dt) {
@@ -32,34 +33,6 @@ class EntitySystem {
         void process(std::function<void(Entity)> func) {
             for (Entity entity : entityManager_.entities_with_components<T...>())
                 func(entity);
-        }
-
-        void calculateSpeed(Entity entity) {
-            auto advanced = entity.component<AdvancedInfo>();
-            advanced->runSpeed_ = 425;
-            if (advanced->moveMode_ == AdvancedInfo::WALK)
-                advanced->runSpeed_ = 200;
-            if (advanced->moveMode_ == AdvancedInfo::RUN) {
-                auto equipped = entity.component<EquippedItems>();
-                auto stats = entity.component<Stats>();
-                advanced->runSpeed_ += stats->dex_ * .8500001;
-                if (equipped->items_[EquippedItems::BOOTS].wearable_.id_ && equipped->items_[EquippedItems::BOOTS].wearable_.life_) {
-                    uint16_t realSpeed = equipped->items_[EquippedItems::BOOTS].runSpeed_ - 65;
-                    uint16_t realMod = realSpeed / 5;
-                    uint16_t dexMod = realMod * (stats->dex_ + realMod) / 23.222;
-                    advanced->runSpeed_ += (realSpeed * 5) + realMod + dexMod;
-                }
-                // TODO : cloth stats
-                // TODO : buffs
-                // TODO : fairy
-                // TODO : dash_up/down
-            } else {
-                // TODO : ride speed
-            }
-        }
-
-        void calculateAtkSpeed(Entity entity) {
-            (void)entity;
         }
 
     private:
