+++ src/map/src/entitysystem.cpp
@@ -97,7 +97,7 @@ Entity EntitySystem::loadCharacter(uint32_t charId, bool platinium) {
     entity.assign<BulletItems>();
     // TODO : write the inventory code
     entity.assign<Inventory>();
-    calculateSpeed(entity);
+    get<UpdateComponents>().calculateSpeed(entity);
     return entity;
 }
